+++ src/admin/controller/base.js
@@ -14,7 +14,7 @@ export default class extends think.controller.base {
 
     // 如果某些 action 不用进行前置操作处理，则在此直接 return; 即可。
     // 例如登录操作
-    if (http.controller === 'login' && http.action === 'login') {
+    if (http.controller === 'user' && http.action === 'login') {
       return;
     }
